packages/midas_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "midas-pipeline"
version = "0.1.0a0"
description = "Unified MIDAS HEDM orchestrator. FF is the single-scan degeneracy of PF; one orchestrator, --scan-mode {ff,pf}."
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.9"
authors = [
    {name = "Hemant Sharma", email = "hsharma@anl.gov"},
]
keywords = [
    "MIDAS", "HEDM", "FF-HEDM", "PF-HEDM", "scanning",
    "diffraction", "pipeline", "PyTorch", "multi-detector",
    "crystallography", "polycrystal", "tomography",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.22",
    "torch>=2.0",
    "zarr>=2.14,<3",
    "h5py>=3.7",
    "tifffile",
    "pandas>=1.5",
    # Sibling MIDAS packages (shared with midas-ff-pipeline).
    "midas-hkls>=0.2.0",
    "midas-peakfit>=0.2.0",
    "midas-transforms>=0.1.1",
    "midas-index>=0.3.0",
    "midas-fit-grain>=0.1.0a0",
    "midas-process-grains>=0.1.0",
    "midas-diffract>=0.1.0",
    "midas-parsl-configs>=0.1.0",
    "midas-stress>=0.6.0",                # canonical orientation/diffraction primitives
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "matplotlib", "ipykernel"]
notebook = ["matplotlib", "ipykernel", "jupyter"]

[project.scripts]
midas-pipeline = "midas_pipeline.cli:main"

[project.urls]
Homepage = "https://github.com/marinerhemant/MIDAS"
Documentation = "https://github.com/marinerhemant/MIDAS"
Issues = "https://github.com/marinerhemant/MIDAS/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["midas_pipeline*"]

[tool.setuptools.package-data]
midas_pipeline = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "slow: full-pipeline regression tests (require golden fixtures)",
    "parity: bit-or-tolerance parity tests vs frozen C-binary output",
    "gpu: tests that require CUDA",
    "mps: tests that require Apple MPS",
]
